initial restructuring of tests
